mt76: add a new queue_ops for cut-through TX arch
In the legacy chips, we move and store the whole packet content to the
on-device memory when we send a TX packet to high speed interface.

However, MT7615 and the following generation use a new TX arch called
'cut-through', which only passes the physical address to the device,
and the device will fetch it the moment (via txp) before it TX packet.

Therefore, this adds a new mt76_queue_ops to satisfy such needs.

Signed-off-by: Ryder Lee <ryder.lee@mediatek.com>
Signed-off-by: Lorenzo Bianconi <lorenzo@kernel.org>